* Add a bunch of LUA- and other missing files to VS8 project file. * Disabled 64 bit compatibility warnings and disabled   comparison between signed and unsigned integer warnings.   (fixing them just makes code uglier because of all the (int) casts) * Fixed all other warnings in VS8 except those in hpiutil2, streflop, lua.   (it's mostly int->boolean performance warnings have been fixed) * Fixed FTBFS when DIRECT_CONTROL_ALLOWED wasn't #defined. * Made it use CrashHandler.cpp on MSVC too (as opposed to CrashReporter).
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@3775 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Rendering/FontTexture.cpp
@@ -472,8 +472,8 @@ Glyph::Glyph(FT_Face& face, int _num) : num(_num), valid(false)
   oyn = oyp - ysize - (2 * stuffing);
 
   // texture coordinates
-  txn = -stuffing;
-  tyn = -stuffing;
+  txn = -(s32)stuffing;
+  tyn = -(s32)stuffing;
   txp = xsize + stuffing;
   typ = ysize + stuffing;
